backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# CORS & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
openpyxl==3.1.2

# Text Processing
nltk==3.8.1
beautifulsoup4==4.12.2
lxml==4.9.3

# PDF Generation
reportlab==4.0.7

# API Requests (for real Keepa integration later)
requests==2.31.0
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
# Apify Client
apify-client==1.7.1